// iot-service/iot-angular/src/sdk/core/models/owner-detail-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ForwardingDatabaseResponseDto } from '../models/forwarding-database-response-dto';
import { ForwardingLogResponseDto } from '../models/forwarding-log-response-dto';
import { ForwardingWebhookResponseDto } from '../models/forwarding-webhook-response-dto';
import { NodeAssignmentSummaryDto } from '../models/node-assignment-summary-dto';
import { OwnerStatisticsDto } from '../models/owner-statistics-dto';
import { ProjectSummaryDto } from '../models/project-summary-dto';
export interface OwnerDetailResponseDto {

  /**
   * Contact person
   */
  contactPerson?: string;

  /**
   * Created timestamp
   */
  createdAt: string;

  /**
   * Database configurations
   */
  forwardingDatabases: Array<ForwardingDatabaseResponseDto>;

  /**
   * Recent forwarding logs
   */
  forwardingLogs: Array<ForwardingLogResponseDto>;

  /**
   * Forwarding settings
   */
  forwardingSettings?: {
};

  /**
   * Webhook configurations
   */
  forwardingWebhooks: Array<ForwardingWebhookResponseDto>;

  /**
   * Owner unique identifier
   */
  idOwner: string;

  /**
   * Industry type
   */
  industry: string;

  /**
   * Owner name
   */
  name: string;

  /**
   * Node assignments
   */
  nodeAssignments: Array<NodeAssignmentSummaryDto>;

  /**
   * List of projects
   */
  projects: Array<ProjectSummaryDto>;

  /**
   * SLA level
   */
  slaLevel?: string;

  /**
   * Owner statistics
   */
  statistics: OwnerStatisticsDto;

  /**
   * Updated timestamp
   */
  updatedAt: string;
}